User Repository Tests

+++ collectMe/src/main/java/learn/collectMe/data/UserJdbcTemplateRepository.java
@@ -28,7 +28,9 @@ public class UserJdbcTemplateRepository implements UserRepository {
     @Override
     public List<User> findAll() {
 
-        final String sql = "select user_id, first_name, last_name, location, username, password_hash, phone, email, enabled  order by last_name;";
+        final String sql = "select user_id, first_name, last_name, location, username, password_hash, phone, email, enabled " +
+                "from user " +
+                "order by last_name;";
 
         return jdbcTemplate.query(sql, new UserWithoutRolesMapper());
     }
@@ -37,7 +39,8 @@ public class UserJdbcTemplateRepository implements UserRepository {
     @Transactional
     public User findById(int userId) {
 
-        final String sql = "select user_id, first_name, last_name, location, username, password_hash, phone, email, enabled "
+         final String sql = "select user_id, first_name, last_name, location, username, password_hash, phone, email, enabled " +
+                "from user "
                 + "where user_id = ?;";
 
         User user = jdbcTemplate.query(sql, new UserWithoutRolesMapper(), userId)
@@ -92,17 +95,20 @@ public class UserJdbcTemplateRepository implements UserRepository {
 
     @Override
     public User add(User user) {
-        final String sql = "insert into user (first_name, last_name, location, phone, email, enabled) "
-                + " values (?,?,?,?,?,?,?);";
+        final String sql = "insert into user (username, first_name, last_name, location, password_hash, phone, email, enabled) "
+                + " values (?,?,?,?,?,?,?,?);";
 
         KeyHolder keyHolder = new GeneratedKeyHolder();
         int rowsAffected = jdbcTemplate.update(connection -> {
             PreparedStatement ps = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
-            ps.setString(1, user.getFirstName());
-            ps.setString(2, user.getLastName());
-            ps.setString(3, user.getLocation());
-            ps.setString(4, user.getPhone());
-            ps.setString(5, user.getEmail());
+            ps.setString(1,user.getUsername());
+            ps.setString(2, user.getFirstName());
+            ps.setString(3, user.getLastName());
+            ps.setString(4, user.getLocation());
+            ps.setString(5, user.getPassword());
+            ps.setString(6, user.getPhone());
+            ps.setString(7, user.getEmail());
+            ps.setBoolean(8, user.isEnabled());
             return ps;
         }, keyHolder);
 
@@ -118,12 +124,19 @@ public class UserJdbcTemplateRepository implements UserRepository {
     @Transactional
     public boolean update(User user) {
         final String sql = "update user set "
-                + "username = ?, "
-                + "enabled = ? "
+                + "username = ?, " +
+                "first_name = ?, " +
+                "last_name = ?, " +
+                "location = ?, " +
+                "password_hash = ?, " +
+                "phone = ?, " +
+                "email = ?, " +
+                "enabled = ? "
                 + "where user_id = ?";
 
         boolean updated = jdbcTemplate.update(sql,
-                user.getUsername(), user.isEnabled(), user.getUserId()) > 0;
+                user.getUsername(), user.getFirstName(), user.getLastName(), user.getLocation(),
+                user.getPassword(), user.getPhone(), user.getEmail(), user.isEnabled(), user.getUserId()) > 0;
 
         if (updated) {
             updateRoles(user);
@@ -134,13 +147,22 @@ public class UserJdbcTemplateRepository implements UserRepository {
 
     @Override
     public boolean deleteById(int userId) {
-        final String sql = "delete from user where user_id = ?";
+        int commentCount = jdbcTemplate.queryForObject(
+                "select count(*) from comment where user_id = ?;", Integer.class, userId);
+        int itemCount = jdbcTemplate.queryForObject(
+                "select count(*) from item where user_id = ?;", Integer.class, userId);
+
+        if ((commentCount > 0) || (itemCount > 0)) {
+            return false;
+        }
+
+        String sql = "delete from user where user_id = ?";
         return jdbcTemplate.update(sql, userId) > 0;
     }
 
     private void addItems(User user) {
 
-        String sql = "select item_id, `name`, description, user_id " +
+        String sql = "select item_id, `name`, description, value, user_id " +
                 "from item where user_id = ?;";
 
         var userItems = jdbcTemplate.query(sql, new ItemMapper(), user.getUserId());
@@ -158,17 +180,17 @@ public class UserJdbcTemplateRepository implements UserRepository {
 
         for (GrantedAuthority role : authorities) {
             String sql = "insert into user_role (user_id, role_id) "
-                    + "select ?, role_id from role where `name` = ?;";
+                    + "select ?, role_id from role where role_name = ?;";
             jdbcTemplate.update(sql, user.getUserId(), role.getAuthority());
         }
     }
 
-    private List<String> getRolesByUsername(String username) {
-        final String sql = "select r.name "
+    public List<String> getRolesByUsername(String username) {
+        final String sql = "select r.role_name "
                 + "from user_role ur "
                 + "inner join role r on ur.role_id = r.role_id "
                 + "inner join user u on ur.user_id = u.user_id "
                 + "where u.username = ?";
-        return jdbcTemplate.query(sql, (rs, rowId) -> rs.getString("name"), username);
+        return jdbcTemplate.query(sql, (rs, rowId) -> rs.getString("role_name"), username);
     }
 }
